requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dependency-injector==4.43.0
deprecated==1.2.15
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
importlib-metadata==8.5.0
jinja2==3.1.4
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
opentelemetry-api==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-sqlalchemy==0.49b2
opentelemetry-semantic-conventions==0.49b2
packaging==24.2
pydantic==2.10.2
pydantic-core==2.27.1
pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.19
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typer==0.14.0
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
wrapt==1.17.0
zipp==3.21.0
